app/requirements.txt
flask
flask-jwt-extended
flask-cors
openai
openpyxl
pymongo
python-dotenv
